+++ src/actions/apply/terraform.ts
@@ -3,12 +3,12 @@ import * as github from "@actions/github";
 import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
-import * as lib from "../../../lib";
-import * as drift from "../../../lib/drift";
-import * as env from "../../../lib/env";
-import * as input from "../../../lib/input";
-import * as aqua from "../../../aqua";
-import * as getTargetConfig from "../../get-target-config";
+import * as lib from "../../lib";
+import * as drift from "../../lib/drift";
+import * as env from "../../lib/env";
+import * as input from "../../lib/input";
+import * as aqua from "../../aqua";
+import * as getTargetConfig from "../get-target-config";
 import * as updateBranchAction from "@csm-actions/update-branch-action";
 import * as githubAppToken from "@suzuki-shunsuke/github-app-token";
 import {
@@ -16,7 +16,7 @@ import {
   FindOptions,
   DownloadArtifactOptions,
 } from "@actions/artifact";
-import * as run from "../run";
+import * as run from "./run";
 
 type WorkflowRun = {
   headSha: string;
